pyproject.toml
[tool.poetry]
name = "lushcloset-backend"
version = "0.1.0"
description = "Backend logic for LushCloset."
authors = ["Chehan Ratnasiri <chehan.rat@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8, <3.9"
django = "^3.0.4"
djangorestframework = "^3.11.0"
djangorestframework-simplejwt = "^4.4.0"
django-cors-headers = "^3.2.1"
psycopg2 = "^2.8.4"
django-filter = "^2.2.0"
googlemaps = "^4.2.0"
boto3 = "^1.12.43"
django-storages = "^1.9.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
